split into multi files for app_manager
