+++ riffle-server/src/app_manager.rs
@@ -15,6 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
+pub mod app;
+pub mod app_configs;
+pub mod partition_identifier;
+mod partition_meta;
+pub mod purge_event;
+pub mod request_context;
+
 use crate::config::{Config, HistoricalAppStoreConfig, StorageType};
 use crate::error::WorkerError;
 use crate::metric::{
@@ -45,6 +52,9 @@ use std::hash::{BuildHasherDefault, Hash, Hasher};
 use std::ops::Deref;
 use std::str::FromStr;
 
+use crate::app_manager::app::App;
+use crate::app_manager::app_configs::AppConfigOptions;
+use crate::app_manager::purge_event::{PurgeEvent, PurgeReason};
 use crate::await_tree::AWAIT_TREE_REGISTRY;
 use crate::block_id_manager::{get_block_id_manager, BlockIdManager};
 use crate::config_reconfigure::ReconfigurableConfManager;
@@ -75,740 +85,6 @@ pub static SHUFFLE_SERVER_ID: OnceLock<String> = OnceLock::new();
 pub static SHUFFLE_SERVER_IP: OnceLock<String> = OnceLock::new();
 
 pub static APP_MANAGER_REF: OnceCell<AppManagerRef> = OnceCell::new();
-
-#[derive(Debug, Clone)]
-pub enum DataDistribution {
-    NORMAL,
-    #[allow(non_camel_case_types)]
-    LOCAL_ORDER,
-}
-
-pub const MAX_CONCURRENCY_PER_PARTITION_TO_WRITE: i32 = 20;
-
-#[derive(Debug, Clone)]
-pub struct AppConfigOptions {
-    pub data_distribution: DataDistribution,
-    pub max_concurrency_per_partition_to_write: i32,
-    pub remote_storage_config_option: Option<RemoteStorageConfig>,
-}
-
-impl AppConfigOptions {
-    pub fn new(
-        data_distribution: DataDistribution,
-        max_concurrency_per_partition_to_write: i32,
-        remote_storage_config_option: Option<RemoteStorageConfig>,
-    ) -> Self {
-        Self {
-            data_distribution,
-            max_concurrency_per_partition_to_write,
-            remote_storage_config_option,
-        }
-    }
-}
-
-impl Default for AppConfigOptions {
-    fn default() -> Self {
-        AppConfigOptions {
-            data_distribution: DataDistribution::LOCAL_ORDER,
-            max_concurrency_per_partition_to_write: 20,
-            remote_storage_config_option: None,
-        }
-    }
-}
-
-// =============================================================
-
-#[derive(Clone, Debug)]
-pub struct RemoteStorageConfig {
-    pub root: String,
-    pub configs: HashMap<String, String>,
-}
-
-impl From<RemoteStorage> for RemoteStorageConfig {
-    fn from(remote_conf: RemoteStorage) -> Self {
-        let root = remote_conf.path;
-        let mut confs = HashMap::new();
-        for kv in remote_conf.remote_storage_conf {
-            confs.insert(kv.key, kv.value);
-        }
-
-        Self {
-            root,
-            configs: confs,
-        }
-    }
-}
-
-// =============================================================
-
-pub struct App {
-    pub app_id: String,
-    app_config_options: AppConfigOptions,
-    latest_heartbeat_time: AtomicU64,
-    store: Arc<HybridStore>,
-
-    memory_capacity: u64,
-
-    // partition limitation
-    partition_limit_enable: bool,
-    partition_limit_threshold: ConfigOption<ByteString>,
-    partition_limit_mem_backpressure_ratio: ConfigOption<f64>,
-
-    total_received_data_size: AtomicU64,
-    total_resident_data_size: AtomicU64,
-
-    // when exceeding the partition-limit-threshold, it will be marked as huge partition
-    huge_partition_number: AtomicU64,
-
-    pub(crate) start_timestamp: u128,
-
-    // key: shuffle_id, val: shuffle's all block_ids bitmap
-    block_id_manager: Arc<Box<dyn BlockIdManager>>,
-
-    // key: (shuffle_id, partition_id)
-    partition_meta_infos: DashMapExtend<(i32, i32), PartitionedMeta, BuildHasherDefault<FxHasher>>,
-
-    // partition split
-    partition_split_enable: bool,
-    partition_split_threshold: ConfigOption<ByteString>,
-
-    // reconfiguration manager
-    reconf_manager: ReconfigurableConfManager,
-}
-
-#[derive(Clone)]
-struct PartitionedMeta {
-    inner: Arc<RwLock<PartitionedMetaInner>>,
-}
-
-struct PartitionedMetaInner {
-    total_size: u64,
-    is_huge_partition: bool,
-
-    is_split: bool,
-}
-
-impl PartitionedMeta {
-    fn new() -> Self {
-        PartitionedMeta {
-            inner: Arc::new(RwLock::new(PartitionedMetaInner {
-                total_size: 0,
-                is_huge_partition: false,
-                is_split: false,
-            })),
-        }
-    }
-
-    fn is_split(&self, uid: &PartitionedUId, threshold: u64) -> Result<bool> {
-        let mut meta = self.inner.write();
-        if meta.is_split {
-            return Ok(true);
-        }
-
-        if (meta.total_size > threshold) {
-            meta.is_split = true;
-            warn!(
-                "Split partition(actual/threshold: {}/{}) for app:{}. shuffle_id:{}. partition_id:{}",
-                meta.total_size, threshold, uid.app_id, uid.shuffle_id, uid.partition_id
-            );
-            return Ok(true);
-        }
-
-        Ok(false)
-    }
-
-    fn get_size(&self) -> Result<u64> {
-        let meta = self.inner.read();
-        Ok(meta.total_size)
-    }
-
-    fn inc_size(&mut self, data_size: i32) -> Result<()> {
-        let mut meta = self.inner.write();
-        meta.total_size += data_size as u64;
-        Ok(())
-    }
-
-    fn is_huge_partition(&self) -> bool {
-        self.inner.read().is_huge_partition
-    }
-
-    fn mark_as_huge_partition(&mut self) {
-        let mut meta = self.inner.write();
-        meta.is_huge_partition = true
-    }
-}
-
-impl App {
-    fn from(
-        app_id: String,
-        config_options: AppConfigOptions,
-        store: Arc<HybridStore>,
-        runtime_manager: RuntimeManager,
-        config: &Config,
-        reconf_manager: &ReconfigurableConfManager,
-    ) -> Self {
-        // todo: should throw exception if register failed.
-        let copy_app_id = app_id.to_string();
-        let app_options = config_options.clone();
-        match store.register_app(RegisterAppContext {
-            app_id: copy_app_id,
-            app_config_options: app_options,
-        }) {
-            Err(error) => {
-                error!("Errors on registering app to store: {:#?}", error,);
-            }
-            _ => {}
-        }
-
-        let memory_capacity =
-            util::parse_raw_to_bytesize(&config.memory_store.as_ref().unwrap().capacity);
-
-        let partition_limit_enable = config.app_config.partition_limit_enable;
-        let partition_limit_threshold: ConfigOption<ByteString> = reconf_manager
-            .register("app_config.partition_limit_threshold")
-            .unwrap();
-        let partition_limit_mem_backpressure_ratio = reconf_manager
-            .register("app_config.partition_limit_memory_backpressure_ratio")
-            .unwrap();
-
-        let partition_split_enable = config.app_config.partition_split_enable;
-        let partition_split_threshold: ConfigOption<ByteString> = reconf_manager
-            .register("app_config.partition_split_threshold")
-            .unwrap();
-
-        // pre-check partition features values.
-        // the partition-split threshold should always be less than the partition-limit threshold
-        if partition_limit_enable && partition_split_enable {
-            if partition_split_threshold.get().as_u64() >= partition_limit_threshold.get().as_u64()
-            {
-                panic!("The value of partition-split threshold should always be less than the partition-limit threshold value!")
-            }
-        }
-
-        let block_id_manager = get_block_id_manager(&config.app_config.block_id_manager_type);
-
-        info!("App=[{}]. block_manager_type: {}. partition_limit/threshold/ratio: {}/{}/{}. partition_split/threshold: {}/{}",
-                &app_id, &config.app_config.block_id_manager_type,
-                partition_limit_enable, partition_limit_threshold.get(), partition_limit_mem_backpressure_ratio.get(),
-                partition_split_enable, partition_split_threshold.get());
-
-        App {
-            app_id,
-            app_config_options: config_options,
-            latest_heartbeat_time: AtomicU64::new(now_timestamp_as_sec()),
-            store,
-            memory_capacity,
-            partition_limit_enable,
-            partition_limit_threshold,
-            partition_limit_mem_backpressure_ratio,
-            partition_meta_infos: DashMapExtend::<
-                (i32, i32),
-                PartitionedMeta,
-                BuildHasherDefault<FxHasher>,
-            >::new(),
-            total_received_data_size: Default::default(),
-            total_resident_data_size: Default::default(),
-            huge_partition_number: Default::default(),
-            start_timestamp: now_timestamp_as_millis(),
-            block_id_manager,
-            partition_split_enable,
-            partition_split_threshold,
-            reconf_manager: reconf_manager.clone(),
-        }
-    }
-
-    pub fn reported_block_id_number(&self) -> u64 {
-        self.block_id_manager.get_blocks_number().unwrap_or(0)
-    }
-
-    pub fn huge_partition_number(&self) -> u64 {
-        self.huge_partition_number.load(SeqCst)
-    }
-
-    pub fn partition_number(&self) -> usize {
-        self.partition_meta_infos.len()
-    }
-
-    fn get_latest_heartbeat_time(&self) -> u64 {
-        self.latest_heartbeat_time.load(SeqCst)
-    }
-
-    pub fn heartbeat(&self) -> Result<()> {
-        let timestamp = now_timestamp_as_sec();
-        self.latest_heartbeat_time.store(timestamp, SeqCst);
-        Ok(())
-    }
-
-    pub fn register_shuffle(&self, shuffle_id: i32) -> Result<()> {
-        self.heartbeat()?;
-        Ok(())
-    }
-
-    pub async fn insert(&self, ctx: WritingViewContext) -> Result<i32, WorkerError> {
-        self.heartbeat()?;
-
-        let len: u64 = ctx.data_size;
-        TOTAL_RECEIVED_DATA.inc_by(len);
-
-        // add the partition size into the meta
-        self.inc_partition_size(&ctx.uid, len)?;
-
-        self.total_received_data_size.fetch_add(len, SeqCst);
-        self.total_resident_data_size.fetch_add(len, SeqCst);
-
-        RESIDENT_BYTES.add(len as i64);
-
-        self.store.insert(ctx).await?;
-        Ok(len as i32)
-    }
-
-    pub async fn select(&self, ctx: ReadingViewContext) -> Result<ResponseData, WorkerError> {
-        self.heartbeat()?;
-
-        let response = self.store.get(ctx).await;
-        response.map(|data| {
-            match &data {
-                ResponseData::Local(local_data) => {
-                    let length = local_data.data.len() as u64;
-                    TOTAL_READ_DATA_FROM_LOCALFILE.inc_by(length);
-                    TOTAL_READ_DATA.inc_by(length);
-                }
-                ResponseData::Mem(mem_data) => {
-                    let length = mem_data.data.len() as u64;
-                    TOTAL_READ_DATA_FROM_MEMORY.inc_by(length);
-                    TOTAL_READ_DATA.inc_by(length);
-                }
-            };
-
-            data
-        })
-    }
-
-    pub async fn list_index(
-        &self,
-        ctx: ReadingIndexViewContext,
-    ) -> Result<ResponseDataIndex, WorkerError> {
-        self.heartbeat()?;
-
-        let response = self.store.get_index(ctx).await;
-        response.map(|data| {
-            match &data {
-                ResponseDataIndex::Local(local_data) => {
-                    let len = local_data.index_data.len();
-                    TOTAL_READ_INDEX_FROM_LOCALFILE.inc_by(len as u64);
-                    TOTAL_READ_DATA.inc_by(len as u64);
-                }
-                _ => {}
-            };
-            data
-        })
-    }
-
-    // Only for test case
-    pub fn mark_huge_partition(&self, uid: &PartitionedUId) -> Result<()> {
-        let mut meta = self.get_partition_meta(uid);
-        meta.mark_as_huge_partition();
-        Ok(())
-    }
-
-    pub fn is_huge_partition(&self, uid: &PartitionedUId) -> Result<bool> {
-        // always mark false when partition limit is not enabled
-        if !self.partition_limit_enable {
-            return Ok(false);
-        }
-
-        let partition_limit_threshold = self.partition_limit_threshold.get().as_u64();
-        let mut meta = self.get_partition_meta(uid);
-        if meta.is_huge_partition() {
-            Ok(true)
-        } else {
-            let data_size = meta.get_size()?;
-            if data_size > partition_limit_threshold {
-                meta.mark_as_huge_partition();
-                self.add_huge_partition_metric();
-                warn!("Partition is marked as huge partition. uid: {:?}", uid);
-                Ok(true)
-            } else {
-                Ok(false)
-            }
-        }
-    }
-
-    fn add_huge_partition_metric(&self) {
-        self.huge_partition_number.fetch_add(1, Ordering::SeqCst);
-        TOTAL_HUGE_PARTITION_NUMBER.inc();
-        GAUGE_HUGE_PARTITION_NUMBER
-            .with_label_values(&[ALL_LABEL])
-            .inc();
-        GAUGE_HUGE_PARTITION_NUMBER
-            .with_label_values(&[self.app_id.as_str()])
-            .inc();
-    }
-
-    fn sub_huge_partition_metric(&self) {
-        let number = self.huge_partition_number.load(SeqCst);
-        if number > 0 {
-            GAUGE_HUGE_PARTITION_NUMBER
-                .with_label_values(&vec![ALL_LABEL])
-                .sub(number as i64);
-
-            if let Err(e) = GAUGE_HUGE_PARTITION_NUMBER.remove_label_values(&[&self.app_id]) {
-                error!(
-                    "Errors on unregistering metric of huge partition number for app:{}. error: {}",
-                    &self.app_id, e
-                )
-            }
-        }
-    }
-
-    pub async fn is_backpressure_of_partition(&self, uid: &PartitionedUId) -> Result<bool> {
-        if !self.is_huge_partition(uid)? {
-            return Ok(false);
-        }
-        let ratio = self.partition_limit_mem_backpressure_ratio.get();
-        let threshold = (self.memory_capacity as f64 * ratio) as u64;
-        let used = self.store.get_memory_buffer_size(uid).await?;
-
-        if used > threshold {
-            info!(
-                "[{:?}] with huge partition (used/limited: {}/{}), it has been limited of writing speed",
-                uid, used, threshold
-            );
-            TOTAL_HUGE_PARTITION_REQUIRE_BUFFER_FAILED.inc();
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
-    pub fn dec_allocated_from_budget(&self, size: i64) -> Result<bool> {
-        self.store.release_allocated_from_hot_store(size)
-    }
-
-    pub fn move_allocated_used_from_budget(&self, size: i64) -> Result<bool> {
-        self.store.move_allocated_to_used_from_hot_store(size)
-    }
-
-    pub async fn require_buffer(
-        &self,
-        ctx: RequireBufferContext,
-    ) -> Result<RequireBufferResponse, WorkerError> {
-        self.heartbeat()?;
-
-        let app_id = &ctx.uid.app_id;
-        let shuffle_id = &ctx.uid.shuffle_id;
-
-        let mut partition_split_candidates = HashSet::new();
-
-        if self.partition_limit_enable || self.partition_split_enable {
-            for partition_id in &ctx.partition_ids {
-                let puid = PartitionedUId::from(app_id.to_owned(), *shuffle_id, *partition_id);
-                let mut split_hit = false;
-
-                // partition split
-                if self.partition_split_enable
-                    && self
-                        .get_partition_meta(&puid)
-                        .is_split(&puid, self.partition_split_threshold.get().into())?
-                {
-                    partition_split_candidates.insert(*partition_id);
-                    split_hit = true;
-                }
-
-                // huge partition limitation
-                if self.is_backpressure_of_partition(&puid).await? {
-                    TOTAL_REQUIRE_BUFFER_FAILED.inc();
-                    return Err(WorkerError::MEMORY_USAGE_LIMITED_BY_HUGE_PARTITION);
-                }
-            }
-        }
-
-        let mut required = self.store.require_buffer(ctx).await.map_err(|err| {
-            TOTAL_REQUIRE_BUFFER_FAILED.inc();
-            err
-        })?;
-        required.split_partitions = partition_split_candidates
-            .iter()
-            .map(|data| *data)
-            .collect();
-        Ok(required)
-    }
-
-    pub async fn release_ticket(&self, ticket_id: i64) -> Result<i64, WorkerError> {
-        self.store
-            .release_ticket(ReleaseTicketContext::from(ticket_id))
-            .await
-    }
-
-    fn get_partition_meta(&self, uid: &PartitionedUId) -> PartitionedMeta {
-        let shuffle_id = uid.shuffle_id;
-        let partition_id = uid.partition_id;
-        self.partition_meta_infos
-            .compute_if_absent((shuffle_id, partition_id), || {
-                TOTAL_PARTITION_NUMBER.inc();
-                GAUGE_PARTITION_NUMBER.inc();
-                PartitionedMeta::new()
-            })
-    }
-
-    pub fn inc_partition_size(&self, uid: &PartitionedUId, size: u64) -> Result<()> {
-        let mut partitioned_meta = self.get_partition_meta(&uid);
-        partitioned_meta.inc_size(size as i32)
-    }
-
-    pub async fn get_multi_block_ids(&self, ctx: GetMultiBlockIdsContext) -> Result<Bytes> {
-        self.heartbeat()?;
-        self.block_id_manager.get_multi_block_ids(ctx).await
-    }
-
-    pub async fn report_multi_block_ids(&self, ctx: ReportMultiBlockIdsContext) -> Result<()> {
-        self.heartbeat()?;
-        let number = self.block_id_manager.report_multi_block_ids(ctx).await?;
-        BLOCK_ID_NUMBER.add(number as i64);
-        Ok(())
-    }
-
-    pub async fn dump_all_huge_partitions_size(&self) -> Result<Vec<u64>> {
-        let mut records = vec![];
-        let view = self.partition_meta_infos.clone().into_read_only();
-        for entry in view.iter() {
-            let val = entry.1;
-            if val.is_huge_partition() {
-                let size = val.get_size()?;
-                records.push(size);
-            }
-        }
-        Ok(records)
-    }
-
-    pub async fn purge(&self, reason: &PurgeReason) -> Result<()> {
-        let (app_id, shuffle_id) = reason.extract();
-        let removed_size = self.store.purge(&PurgeDataContext::new(reason)).await?;
-        self.total_resident_data_size
-            .fetch_sub(removed_size as u64, SeqCst);
-
-        RESIDENT_BYTES.sub(removed_size);
-
-        if let Some(shuffle_id) = shuffle_id {
-            // shuffle level bitmap deletion
-            let purged_number = self.block_id_manager.purge_block_ids(shuffle_id).await?;
-            BLOCK_ID_NUMBER.sub(purged_number as i64);
-
-            let mut deletion_keys = vec![];
-            let view = self.partition_meta_infos.clone().into_read_only();
-            for entry in view.iter() {
-                let key = entry.0;
-                if shuffle_id == key.0 {
-                    deletion_keys.push(key);
-                }
-            }
-            GAUGE_PARTITION_NUMBER.sub(deletion_keys.len() as i64);
-            let mut huge_partition_cnt = 0;
-            for deletion_key in deletion_keys {
-                if let Some(meta) = self.partition_meta_infos.remove(deletion_key) {
-                    if meta.1.is_huge_partition() {
-                        huge_partition_cnt += 1;
-                    }
-                }
-            }
-            GAUGE_HUGE_PARTITION_NUMBER
-                .with_label_values(&vec![ALL_LABEL])
-                .sub(huge_partition_cnt as i64);
-        } else {
-            // app level deletion
-            GAUGE_PARTITION_NUMBER.sub(self.partition_meta_infos.len() as i64);
-            self.sub_huge_partition_metric();
-
-            BLOCK_ID_NUMBER.sub(self.block_id_manager.get_blocks_number()? as i64);
-        }
-
-        Ok(())
-    }
-
-    pub fn total_received_data_size(&self) -> u64 {
-        self.total_received_data_size.load(SeqCst)
-    }
-
-    pub fn total_resident_data_size(&self) -> u64 {
-        self.total_resident_data_size.load(SeqCst)
-    }
-}
-
-#[allow(non_camel_case_types)]
-#[derive(Debug, Clone)]
-pub enum PurgeReason {
-    SHUFFLE_LEVEL_EXPLICIT_UNREGISTER(String, i32),
-    APP_LEVEL_EXPLICIT_UNREGISTER(String),
-    APP_LEVEL_HEARTBEAT_TIMEOUT(String),
-}
-
-impl PurgeReason {
-    pub fn extract(&self) -> (String, Option<i32>) {
-        match &self {
-            PurgeReason::SHUFFLE_LEVEL_EXPLICIT_UNREGISTER(x, y) => (x.to_owned(), Some(*y)),
-            PurgeReason::APP_LEVEL_EXPLICIT_UNREGISTER(x) => (x.to_owned(), None),
-            PurgeReason::APP_LEVEL_HEARTBEAT_TIMEOUT(x) => (x.to_owned(), None),
-        }
-    }
-
-    pub fn extract_app_id(&self) -> String {
-        match &self {
-            PurgeReason::SHUFFLE_LEVEL_EXPLICIT_UNREGISTER(x, y) => x.to_owned(),
-            PurgeReason::APP_LEVEL_EXPLICIT_UNREGISTER(x) => x.to_owned(),
-            PurgeReason::APP_LEVEL_HEARTBEAT_TIMEOUT(x) => x.to_owned(),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct PurgeDataContext {
-    pub purge_reason: PurgeReason,
-}
-
-impl PurgeDataContext {
-    pub fn new(reason: &PurgeReason) -> PurgeDataContext {
-        PurgeDataContext {
-            purge_reason: reason.clone(),
-        }
-    }
-}
-
-impl Deref for PurgeDataContext {
-    type Target = PurgeReason;
-
-    fn deref(&self) -> &Self::Target {
-        &self.purge_reason
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ReportBlocksContext {
-    pub(crate) uid: PartitionedUId,
-    pub(crate) blocks: Vec<i64>,
-}
-
-#[derive(Debug, Clone)]
-pub struct ReportMultiBlockIdsContext {
-    pub shuffle_id: i32,
-    pub block_ids: HashMap<i32, Vec<i64>>,
-}
-impl ReportMultiBlockIdsContext {
-    pub fn new(shuffle_id: i32, block_ids: HashMap<i32, Vec<i64>>) -> ReportMultiBlockIdsContext {
-        Self {
-            shuffle_id,
-            block_ids,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct GetMultiBlockIdsContext {
-    pub shuffle_id: i32,
-    pub partition_ids: Vec<i32>,
-    pub layout: IdLayout,
-}
-
-#[derive(Debug, Clone)]
-pub struct GetBlocksContext {
-    pub(crate) uid: PartitionedUId,
-}
-
-#[derive(Debug, Clone)]
-pub struct WritingViewContext {
-    pub uid: PartitionedUId,
-    pub data_blocks: Vec<Block>,
-    pub data_size: u64,
-}
-
-impl WritingViewContext {
-    // only for test
-    pub fn create_for_test(uid: PartitionedUId, data_blocks: Vec<Block>) -> Self {
-        WritingViewContext {
-            uid,
-            data_blocks,
-            data_size: 0,
-        }
-    }
-
-    // only for test
-    pub fn new_with_size(uid: PartitionedUId, data_blocks: Vec<Block>, data_size: u64) -> Self {
-        WritingViewContext {
-            uid,
-            data_blocks,
-            data_size,
-        }
-    }
-
-    pub fn new(uid: PartitionedUId, data_blocks: Vec<Block>) -> Self {
-        let len: u64 = data_blocks.iter().map(|block| block.length).sum::<i32>() as u64;
-        WritingViewContext {
-            uid,
-            data_blocks,
-            data_size: len,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ReadingViewContext {
-    pub uid: PartitionedUId,
-    pub reading_options: ReadingOptions,
-    pub serialized_expected_task_ids_bitmap: Option<Treemap>,
-}
-
-pub struct ReadingIndexViewContext {
-    pub partition_id: PartitionedUId,
-}
-
-#[derive(Debug, Clone)]
-pub struct RequireBufferContext {
-    pub uid: PartitionedUId,
-    pub size: i64,
-    // todo: we should replace uid with (app_id, shuffle_id).
-    pub partition_ids: Vec<i32>,
-}
-
-#[derive(Debug, Clone)]
-pub struct RegisterAppContext {
-    pub app_id: String,
-    pub app_config_options: AppConfigOptions,
-}
-
-#[derive(Debug, Clone)]
-pub struct ReleaseTicketContext {
-    pub(crate) ticket_id: i64,
-}
-
-impl From<i64> for ReleaseTicketContext {
-    fn from(value: i64) -> Self {
-        Self { ticket_id: value }
-    }
-}
-
-impl RequireBufferContext {
-    pub fn create_for_test(uid: PartitionedUId, size: i64) -> Self {
-        Self {
-            uid,
-            size,
-            partition_ids: vec![],
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum ReadingOptions {
-    #[allow(non_camel_case_types)]
-    MEMORY_LAST_BLOCK_ID_AND_MAX_SIZE(i64, i64),
-    #[allow(non_camel_case_types)]
-    FILE_OFFSET_AND_LEN(i64, i64),
-}
-
-// ==========================================================
-
-#[derive(Debug, Clone)]
-pub struct PurgeEvent {
-    reason: PurgeReason,
-}
 
 pub type AppManagerRef = Arc<AppManager>;
 
@@ -1085,39 +361,14 @@ impl AppManager {
     }
 }
 
-#[derive(Ord, PartialOrd, Eq, PartialEq, Default, Debug, Hash, Clone)]
-pub struct PartitionedUId {
-    pub app_id: String,
-    pub shuffle_id: i32,
-    pub partition_id: i32,
-}
-
-impl PartitionedUId {
-    pub fn from(app_id: String, shuffle_id: i32, partition_id: i32) -> PartitionedUId {
-        PartitionedUId {
-            app_id,
-            shuffle_id,
-            partition_id,
-        }
-    }
-
-    pub fn get_hash(uid: &PartitionedUId) -> u64 {
-        let mut hasher = DefaultHasher::new();
-
-        uid.hash(&mut hasher);
-        let hash_value = hasher.finish();
-
-        hash_value
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod test {
-    use crate::app_manager::{
-        AppManager, GetBlocksContext, GetMultiBlockIdsContext, PartitionedUId, PurgeReason,
-        ReadingOptions, ReadingViewContext, ReportBlocksContext, ReportMultiBlockIdsContext,
+    use crate::app_manager::partition_identifier::PartitionedUId;
+    use crate::app_manager::request_context::{
+        GetMultiBlockIdsContext, ReadingOptions, ReadingViewContext, ReportMultiBlockIdsContext,
         RequireBufferContext, WritingViewContext,
     };
+    use crate::app_manager::{AppManager, PurgeReason};
     use crate::config::{Config, HybridStoreConfig, LocalfileStoreConfig, MemoryStoreConfig};
     use crate::config_reconfigure::ReconfigurableConfManager;
     use crate::error::WorkerError;
