+++ riffle-server/src/http/apps.rs
@@ -1,4 +1,5 @@
-use crate::app_manager::{self, App, APP_MANAGER_REF};
+use crate::app_manager::app::App;
+use crate::app_manager::{self, APP_MANAGER_REF};
 use crate::http::{Format, Handler};
 use crate::util;
 use chrono::{Local, TimeZone, Utc};
