+++ riffle-server/src/store/hybrid.rs
@@ -15,10 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use crate::app_manager::{
-    AppManagerRef, PartitionedUId, PurgeDataContext, ReadingIndexViewContext, ReadingOptions,
-    ReadingViewContext, RegisterAppContext, ReleaseTicketContext, RequireBufferContext,
-    WritingViewContext,
+use crate::app_manager::request_context::{
+    PurgeDataContext, ReadingIndexViewContext, ReadingOptions, ReadingViewContext,
+    RegisterAppContext, ReleaseTicketContext, RequireBufferContext, WritingViewContext,
 };
 
 use crate::config::{Config, HybridStoreConfig, StorageType};
@@ -56,6 +55,8 @@ use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::Mutex;
 
+use crate::app_manager::partition_identifier::PartitionedUId;
+use crate::app_manager::AppManagerRef;
 use crate::config_reconfigure::ReconfigurableConfManager;
 use crate::runtime::manager::RuntimeManager;
 use crate::store::local::LocalfileStoreStat;
@@ -741,10 +742,9 @@ impl Store for HybridStore {
 
 #[cfg(test)]
 pub(crate) mod tests {
-    use crate::app_manager::ReadingOptions::MEMORY_LAST_BLOCK_ID_AND_MAX_SIZE;
-    use crate::app_manager::{
-        PartitionedUId, ReadingIndexViewContext, ReadingOptions, ReadingViewContext,
-        WritingViewContext,
+    use crate::app_manager::request_context::ReadingOptions::MEMORY_LAST_BLOCK_ID_AND_MAX_SIZE;
+    use crate::app_manager::request_context::{
+        ReadingIndexViewContext, ReadingOptions, ReadingViewContext, WritingViewContext,
     };
     use crate::config::{
         Config, HybridStoreConfig, LocalfileStoreConfig, MemoryStoreConfig, StorageType,
@@ -762,6 +762,7 @@ pub(crate) mod tests {
     use std::sync::Arc;
     use std::thread;
 
+    use crate::app_manager::partition_identifier::PartitionedUId;
     use crate::config_reconfigure::ReconfigurableConfManager;
     use serde::de::Unexpected::Seq;
     use std::time::Duration;
